+++ application/views/fleet.php
@@ -1,0 +1,25 @@
+<h1>Fleet</h1>
+<br /><br />
+
+<table>
+    <tr>
+        <td>
+            <u>ID</u>
+        </td>
+        <td>
+            <u>Make</u>
+        </td>
+        <td>
+            <u>Model</u>
+        </td>
+    </tr>  
+    {fleet}
+    <tr>
+        <td>
+            <a href="/fleet/{id}">{id}</a>
+        </td>
+        <td>{make}</td>
+        <td>{model}</td>
+    </tr>
+    {/fleet}
+</table>